requirements.txt
boto3>=1.9.36
avro-python3>=1.8.2
